Rewrite to generic push-only smb proxy
Can be run without CAP_SYS_ADMIN but is more error prone if the samba share fails.
We maybe need a fallback method for uploading archive files and/or callback to the client in case the upload failed.
This could maybe be circumvented by passing through the folder contents of the container host directly.